requirements.txt
discord.py[voice]==2.6.4
ffmpeg-python
gtts==2.5.4
pycryptodome==3.23.0
validators==0.35.0
yt_dlp==2025.12.8
